Use PowerShell verbs in DBUtils
In the DBUtils-Module some functions have unaproved verbs. To follow
PowerShell best practive and to avoid a Warning when importing the Module,
only aproved verbs should be used for advanced functions.
I renamed the functions so that all use approved verbs and created aliases
with the old name to not break compatibility which existing scripts.